Move Logger to baaahs.util.
