+++ src/commonMain/kotlin/baaahs/Pinky.kt
@@ -24,6 +24,7 @@ import baaahs.proto.*
 import baaahs.shaders.PixelBrainShader
 import baaahs.show.Show
 import baaahs.util.Framerate
+import baaahs.util.Logger
 import kotlinx.coroutines.*
 import kotlinx.serialization.Serializable
 import kotlin.coroutines.CoroutineContext
