+++ src/commonMain/kotlin/baaahs/api/ws/WebSocketClient.kt
@@ -1,11 +1,11 @@
 package baaahs.api.ws
 
-import baaahs.Logger
 import baaahs.imaging.Bitmap
 import baaahs.mapper.MappingSession
 import baaahs.mapper.Storage
 import baaahs.net.Network
 import baaahs.proto.Ports
+import baaahs.util.Logger
 import com.soywiz.klock.DateTime
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.MainScope
