+++ src/commonMain/kotlin/baaahs/plugin/Plugins.kt
@@ -1,12 +1,12 @@
 package baaahs.plugin
 
 import baaahs.Gadget
-import baaahs.Logger
 import baaahs.getBang
 import baaahs.gl.patch.ContentType
 import baaahs.gl.shader.InputPort
 import baaahs.show.Control
 import baaahs.show.DataSource
+import baaahs.util.Logger
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonObject
