+++ src/jvmMain/kotlin/baaahs/JvmSoundAnalyzer.kt
@@ -1,5 +1,6 @@
 package baaahs
 
+import baaahs.util.Logger
 import be.tarsos.dsp.AudioDispatcher
 import be.tarsos.dsp.AudioEvent
 import be.tarsos.dsp.AudioProcessor
