+++ src/jvmTest/kotlin/baaahs/show/mutable/EditingShaderSpec.kt
@@ -35,7 +35,7 @@ object EditingShaderSpec : Spek({
         val notifiedStates by value { arrayListOf<State>() }
 
         val editingShader by value {
-            EditingShader(mutableShow, mutableShaderInstance) {
+            EditingShader(mutableShaderInstance) {
                 mockShaderBuilder
             }.also { it.addObserver { notifiedStates.add(it.state) } }
         }
